refactor(#1952,#1953): split pressure_drop_interface and rest_api into focused submodules
Issue #1952 — pressure_drop_interface.py (was ~1400 lines via _pdi_* private modules):
- Add pressure_drop_units.py: _convert_temperature, _convert_pressure
- Add pressure_drop_validation.py: _validate_*, _log_validation_report, validate_inputs
- Add pressure_drop_results.py: _format_results, print_results, _print_*, _generate_recommendations, _wrap_text
- Remove obsolete _pdi_formatters.py, _pdi_helpers.py, _pdi_unit_converters.py, _pdi_validators.py
- Thin pressure_drop_interface.py now re-exports all public names for backward compatibility

Issue #1953 — rest_api.py (1199 lines):
- Add rest_api_types.py: HTTPMethod, APIRequest, APIResponse, Route
- Add rest_api_routes.py: ModelGenerationAPI with all handler methods
- Add rest_api_flask.py: FlaskAdapter
- Add rest_api_fastapi.py: FastAPIAdapter
- Slim rest_api.py is now a backward-compatibility re-export shim

No logic changes. All public names remain importable from original modules.
Update ruff.toml per-file-ignores to track renamed files.
