+++ src/shared/python/model_generation/api/rest_api_fastapi.py
@@ -1,0 +1,81 @@
+"""FastAPI adapter for the model generation REST API.
+
+Registers ``ModelGenerationAPI`` routes with a FastAPI application instance.
+"""
+
+from __future__ import annotations
+
+import logging
+from collections.abc import Callable
+from typing import Any
+
+from .rest_api_routes import ModelGenerationAPI
+from .rest_api_types import APIRequest, HTTPMethod, Route
+
+logger = logging.getLogger(__name__)
+
+
+class FastAPIAdapter:
+    """Adapter for FastAPI framework."""
+
+    def __init__(self, api: ModelGenerationAPI) -> None:
+        if not (api is not None):
+            raise ValueError("api must be provided")
+        self.api = api
+
+    def register(self, app: Any) -> None:
+        """Register routes with FastAPI app."""
+        from fastapi import Request, Response
+        from fastapi.responses import JSONResponse
+
+        for route in self.api.get_routes():
+
+            async def make_handler(r: Route) -> Callable[..., Any]:
+                async def handler(request: Request, **kwargs: Any) -> Any:
+                    body = None
+                    try:
+                        body = await request.json()
+                    except (ValueError, UnicodeDecodeError) as e:
+                        logger.debug("Failed to parse request JSON body: %s", e)
+
+                    files = {}
+                    form = await request.form()
+                    for key, value in form.items():
+                        if hasattr(value, "read"):
+                            files[key] = await value.read()
+
+                    api_request = APIRequest(
+                        method=HTTPMethod(request.method),
+                        path=request.url.path,
+                        query_params={**request.query_params, **kwargs},
+                        body=body,
+                        files=files,
+                        headers=dict(request.headers),
+                    )
+
+                    response = self.api.handle_request(api_request)
+
+                    if isinstance(response.body, bytes):
+                        return Response(
+                            content=response.body,
+                            status_code=response.status_code,
+                            media_type=response.content_type,
+                            headers=response.headers,
+                        )
+                    else:
+                        return JSONResponse(
+                            content=response.body,
+                            status_code=response.status_code,
+                            headers=response.headers,
+                        )
+
+                return handler
+
+            # FastAPI uses {param} format already
+            app.add_api_route(
+                route.path,
+                make_handler(route),
+                methods=[route.method.value],
+                tags=route.tags,
+                summary=route.description,
+            )
